对eas-grademanage进行修复（参数类型）

+++ eams-java/eams-apis/src/main/java/com/zeroone/star/project/j6/eas/GradeManagerApis.java
@@ -23,7 +23,7 @@ public interface GradeManagerApis {
     /**
      * 获取年级详情
      */
-    JsonVO<ClassGradeDto> getGradeById(Integer id);
+    JsonVO<ClassGradeDto> getGradeById(Long id);
 
     /**
      * 获取年级名称列表
